+++ packages/dtfit-experimental/src/dtfit_experimental/experiments/cases/08_gpu_batched_projection/backend.py
@@ -31,9 +31,9 @@ import time
 
 import numpy as np
 
-from dtfit.scale import project_spectra
 from dtfit._core._spectral import make_basis
 from dtfit_experimental import available_backends, resolve_backend
+from dtfit_experimental.scale import project_spectra
 
 from dtfit_experimental.experiments.common import fmt